+++ R/camera.R
@@ -3,10 +3,10 @@
 #' Add a camera to your graph.
 #'
 #' @param sg An object of class \code{sigmajs}as intatiated by \code{\link{sigmajs}}.
-#' @param name Name of camera.
+#' @param elementId Id of graph that initialised the camera.
 #' @param initialise Whether to initialise the camera.
 #'
-#' @note Cameras should only be initialised once, hence the argument.
+#' @note Cameras should only be initialised once.
 #'
 #' @examples
 #' \dontrun{
@@ -15,33 +15,20 @@
 #' 
 #' @rdname camera
 #' @export
-sg_camera <- function(sg, name, initialise = FALSE) {
-	if (missing(name))
-		stop("missing name", call. = FALSE)
-
-	if (missing(sg))
-		stop("missing sg", call. = FALSE)
-
-	if (!inherits(sg, "sigmajs"))
-		stop("sg must be of class sigmajs", call. = FALSE)
-
-	if(isTRUE(initialise))
-		sg$x$cameraInitialise <- TRUE
-	
-	sg$x$camera <- name
-
-	sg
-}
-
-#' @rdname camera
-#' @export
-sg_add_camera <- function(name){
-	if(missing(name))
-		stop("missing name", call. = FALSE)
-	
-	cam <- paste0("s.addCamera('", name, "');")
-
-	shiny::tags$head(
-		shiny::tags$script(cam)
-	)
+sg_camera <- function(sg, elementId = NULL, initialise = FALSE) {
+  if (missing(elementId) && !isTRUE(initialise))
+    stop("must pass element id if not initialising the camera", call. = FALSE)
+  
+  if (missing(sg))
+    stop("missing sg", call. = FALSE)
+  
+  if (!inherits(sg, "sigmajs"))
+    stop("sg must be of class sigmajs", call. = FALSE)
+  
+  sg$x$camera <- list(
+    init = initialise,
+    id = elementId
+  )
+  
+  sg
 }